+++ src/core/src/lib/extensions/field-form/field-form.ts
@@ -7,7 +7,7 @@ import {
   ValidatorFn,
   AsyncValidatorFn,
 } from '@angular/forms';
-import { getFieldValue, defineHiddenProp, hasKey } from '../../utils';
+import { getFieldValue, defineHiddenProp, hasKey, getKeyPath } from '../../utils';
 import { registerControl, findControl, updateValidity } from './utils';
 import { of } from 'rxjs';
 
@@ -107,6 +107,15 @@ export class FieldFormExtension implements FormlyExtension {
 
         if (markForCheck) {
           updateValidity(c, true);
+
+          // update validity of `FormGroup` instance created by field with nested key.
+          let parent = c.parent;
+          for (let i = 1; i < getKeyPath(field).length; i++) {
+            if (parent) {
+              updateValidity(parent, true);
+              parent = parent.parent;
+            }
+          }
         }
       }
     }