+++ src/core/src/lib/templates/formly.attributes.spec.ts
@@ -1,3 +1,4 @@
+import { fakeAsync, tick } from '@angular/core/testing';
 import { FormControl } from '@angular/forms';
 import { FormlyFieldConfig } from '@ngx-formly/core';
 import { createComponent, ɵCustomEvent } from '@ngx-formly/core/testing';
@@ -178,16 +179,18 @@ describe('FormlyAttributes Component', () => {
   });
 
   describe('focus the element', () => {
-    it(`should focus the element when focus is set to "true" and then blurred when it's set to "false"`, () => {
+    it(`should focus the element when focus is set to "true" and then blurred when it's set to "false"`, fakeAsync(() => {
       const { detectChanges, query, field } = renderComponent({ focus: true });
       const inputEl = <HTMLInputElement>query('input').nativeElement;
 
+      tick();
       expect(document.activeElement === inputEl).toBeTrue();
 
       field.focus = false;
+      tick();
       detectChanges();
       expect(document.activeElement === inputEl).toBeFalse();
-    });
+    }));
 
     it('should change field focus when the element is focused or blurred', () => {
       const { query, field } = renderComponent({ focus: false });
@@ -226,7 +229,7 @@ describe('FormlyAttributes Component', () => {
       expect(query('input').attributes.id).toEqual('foo');
     });
 
-    it(`should focus the first element when mutliple formlyAttributes is present`, () => {
+    it(`should focus the first element when mutliple formlyAttributes is present`, fakeAsync(() => {
       const { detectChanges, field, query } = renderComponent(
         { focus: true },
         {
@@ -239,8 +242,9 @@ describe('FormlyAttributes Component', () => {
       );
 
       field.focus = true;
+      tick();
       detectChanges();
       expect(document.activeElement === query('input').nativeElement).toBeTrue();
-    });
+    }));
   });
 });